feat: add password field to delete user endpoint

+++ src/user/user.service.ts
@@ -497,7 +497,7 @@ export class UserService {
     }
   }
 
-  async deleteUser(userId: string): Promise<any> {
+  async deleteUser(userId: string, deleteUserDto: { password: string }): Promise<any> {
     if (!userId) {
       throw new BadRequestException('User ID is required');
     }
@@ -505,6 +505,17 @@ export class UserService {
     try {
       const user = await this.userRepository.findOne({
         where: { id: userId },
+        select: {
+          id: true,
+          password: true,
+          vacancies: true,
+          resumes: true,
+          coverLetters: true,
+          projects: true,
+          notes: true,
+          events: true,
+          predictions: true
+        },
         relations: [
           'vacancies',
           'resumes',
@@ -521,6 +532,11 @@ export class UserService {
           'No account found with this ID',
           HttpStatus.NOT_FOUND,
         );
+      }
+
+      const isPasswordValid = await argon2.verify(user.password, deleteUserDto.password);
+      if (!isPasswordValid) {
+        throw new UnauthorizedException('Invalid password');
       }
 
       await this.userRepository.remove(user);
